requirements.txt
fastapi==0.105.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
uvicorn==0.24.0.post1
python-dotenv==1.0.0
mariadb==1.1.8
cryptography==41.0.7
sqlalchemy==2.0.23